feat(web): implement announcement banner with dismissal functionality
