+++ apps/web/app/_components/announcement-banner-wrapper.tsx
@@ -1,18 +1,17 @@
 import { COOKIE_NAMES } from '@shinju-date/constants'
 import { cookies } from 'next/headers'
 import { Suspense } from 'react'
-import { getAnnouncement } from '@/lib/announcement/get-announcement'
+import { getAnnouncement } from '../_lib/actions'
 import { AnnouncementBanner } from './announcement-banner'
 
 async function AnnouncementContent() {
+  const cookieStore = await cookies()
   const announcement = await getAnnouncement()
 
   if (!announcement) {
     return null
   }
 
-  // Check if the user has dismissed this announcement
-  const cookieStore = await cookies()
   const dismissedId = cookieStore.get(
     COOKIE_NAMES.DISMISSED_ANNOUNCEMENT_ID,
   )?.value
@@ -22,15 +21,7 @@ async function AnnouncementContent() {
     return null
   }
 
-  return (
-    <AnnouncementBanner
-      endAt={announcement.end_at}
-      initialId={announcement.id}
-      level={announcement.level}
-      message={announcement.message}
-      startAt={announcement.start_at}
-    />
-  )
+  return <AnnouncementBanner announcement={announcement} />
 }
 
 export function AnnouncementBannerWrapper() {
