+++ apps/web/app/_lib/actions.ts
@@ -1,7 +1,62 @@
 'use server'
 
+import { COOKIE_NAMES } from '@shinju-date/constants'
+import type { Tables } from '@shinju-date/database'
+import { logger } from '@shinju-date/logger'
+import { toDBString } from '@shinju-date/temporal-fns'
 import { track } from '@vercel/analytics/server'
+import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
+import { Temporal } from 'temporal-polyfill'
+import { supabaseClient } from '@/lib/supabase'
+
+const ONE_DAY_IN_SECONDS = 60 * 60 * 24
+
+export type Announcement = Pick<
+  Tables<'announcements'>,
+  'end_at' | 'id' | 'level' | 'message' | 'start_at'
+>
+
+/**
+ * Get the currently active announcement
+ * Returns null if no announcement is active
+ */
+export async function getAnnouncement(): Promise<Announcement | null> {
+  const now = Temporal.Now.instant()
+
+  const { data, error } = await supabaseClient
+    .from('announcements')
+    .select('id, message, level, start_at, end_at')
+    .eq('enabled', true)
+    .lte('start_at', toDBString(now))
+    .gte('end_at', toDBString(now))
+    .order('created_at', { ascending: false })
+    .maybeSingle()
+
+  if (error) {
+    logger.error('Failed to fetch announcement', { error })
+
+    return null
+  }
+
+  return data
+}
+
+/**
+ * Server action to dismiss an announcement by saving its ID to a cookie
+ * This prevents the announcement from being shown again until the cookie expires
+ */
+export async function dismissAnnouncement(announcementId: string) {
+  const cookieStore = await cookies()
+
+  cookieStore.set(COOKIE_NAMES.DISMISSED_ANNOUNCEMENT_ID, announcementId, {
+    httpOnly: true,
+    maxAge: ONE_DAY_IN_SECONDS,
+    path: '/',
+    sameSite: 'lax',
+    secure: process.env.NODE_ENV === 'production',
+  })
+}
 
 export async function search(formData: FormData) {
   const query = formData.get('query')
