+++ apps/web/app/layout.tsx
@@ -2,10 +2,10 @@ import './globals.css'
 import { Analytics } from '@vercel/analytics/react'
 import type { Metadata, Viewport } from 'next'
 import Link from 'next/link'
-import { AnnouncementBannerWrapper } from '@/components/announcement/announcement-banner-wrapper'
 import PageVisitTracker from '@/components/page-visit-tracker'
 import { TimerProvider } from '@/components/timer'
 import { title as siteName, themeColor } from '@/lib/constants'
+import { AnnouncementBannerWrapper } from './_components/announcement-banner-wrapper'
 import { ContactLink } from './_components/contact-link'
 import { Providers } from './_components/providers'
 import { SearchButton } from './_components/search-button'
